python/python-ai-agent-tutorial/simple-agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mcp-adapters>=0.1.7",
    "langchain-mcp-tools>=0.2.4",
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
    "websockets>=15.0.1",
]
